requirements.txt
numpy
pillow
opencv-python-headless
tifffile
supabase-client